+++ ServerlessComputing/GoogleCloud/requirements.txt
@@ -1,0 +1,21 @@
+astroid==2.3.3
+certifi==2019.11.28
+chardet==3.0.4
+Click==7.0
+Flask==1.1.1
+idna==2.8
+isort==4.3.21
+itsdangerous==1.1.0
+Jinja2==2.10.3
+lazy-object-proxy==1.4.3
+MarkupSafe==1.1.1
+mccabe==0.6.1
+Pillow==6.2.1
+pkg-resources==0.0.0
+pylint==2.4.4
+requests==2.22.0
+six==1.13.0
+typed-ast==1.4.0
+urllib3==1.25.7
+Werkzeug==0.16.0
+wrapt==1.11.2
