--- ServerlessComputing/GoogleCloud/src/requirements.txt
@@ -1,2 +1,0 @@
-pillow
-flask==1.1.1
