+++ ServerlessComputing/GoogleCloud/src/resize/main.py
