+++ ServerlessComputing/GoogleCloud/src/resize/requirements.txt
@@ -1,0 +1,2 @@
+pillow==6.2.1
+flask==1.1.1
