--- ServerlessComputing/GoogleCloud/src_client/requirements.txt
@@ -1,1 +1,0 @@
-google-auth==1.8.2
